refactor(appserver): isolate turn handlers
