+++ internal/appserver/server.go
@@ -3,22 +3,17 @@ package appserver
 import (
 	"bufio"
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"reflect"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/blueberrycongee/wuu/internal/agent"
 	"github.com/blueberrycongee/wuu/internal/providers"
 	"github.com/blueberrycongee/wuu/internal/runtime"
-	"github.com/blueberrycongee/wuu/internal/session"
-	"github.com/blueberrycongee/wuu/internal/subagent"
 )
 
 var errShutdown = errors.New("app-server shutdown requested")
@@ -189,500 +184,10 @@ func (s *Server) handleClientResponse(raw []byte) error {
 	return nil
 }
 
-func (s *Server) handleTurnStart(ctx context.Context, req Request) error {
-	var params TurnStartParams
-	if err := decodeParams(req.Params, &params); err != nil {
-		return s.writeResponse(req.ID, nil, err)
-	}
-	params.ThreadID = strings.TrimSpace(params.ThreadID)
-	params.Prompt = strings.TrimSpace(params.Prompt)
-	if params.ThreadID == "" {
-		return s.writeResponse(req.ID, nil, errors.New("thread_id is required"))
-	}
-	images, err := normalizeTurnStartImages(params.Images)
-	if err != nil {
-		return s.writeResponse(req.ID, nil, err)
-	}
-	if params.Prompt == "" && len(images) == 0 {
-		return s.writeResponse(req.ID, nil, errors.New("prompt or image is required"))
-	}
-	th := s.thread(params.ThreadID)
-	if th == nil {
-		return s.writeResponse(req.ID, nil, fmt.Errorf("thread %q not found", params.ThreadID))
-	}
-	threadRuntime, err := s.ensureThreadRuntime(th)
-	if err != nil {
-		return s.writeResponse(req.ID, nil, err)
-	}
-
-	turnID := session.NewID()
-	turnCtx, cancel := context.WithCancel(ctx)
-	userMsg := providers.ChatMessage{Role: "user", Content: params.Prompt, Images: images}
-	now := time.Now().UTC()
-
-	th.mu.Lock()
-	if th.running {
-		th.mu.Unlock()
-		cancel()
-		return s.writeResponse(req.ID, nil, fmt.Errorf("thread %q already has a running turn", params.ThreadID))
-	}
-	if err := appendChatMessage(th.MemoryPath, userMsg); err != nil {
-		th.mu.Unlock()
-		cancel()
-		return s.writeResponse(req.ID, nil, err)
-	}
-	history := append([]providers.ChatMessage(nil), th.History...)
-	history = append(history, userMsg)
-	th.History = history
-	th.cancel = cancel
-	turn := th.startTurnLocked(turnID, userMsg, now)
-	th.mu.Unlock()
-
-	if err := s.writeResponse(req.ID, TurnStartResult{Turn: turn}, nil); err != nil {
-		cancel()
-		return err
-	}
-	if err := s.writeNotification(NotificationTurnStarted, TurnStartedNotification{
-		ThreadID: params.ThreadID,
-		Turn:     turn,
-	}); err != nil {
-		cancel()
-		return err
-	}
-
-	go s.runTurn(turnCtx, th, threadRuntime, turnID, history)
-	return nil
-}
-
-func (s *Server) ensureThreadRuntime(th *threadState) (*runtime.ThreadRuntime, error) {
-	if th == nil {
-		return nil, errors.New("thread is required")
-	}
-	th.mu.Lock()
-	existing := th.execRuntime
-	history := cloneHistory(th.History)
-	th.mu.Unlock()
-	if existing != nil {
-		return existing, nil
-	}
-	if s.rt == nil {
-		return nil, errors.New("runtime session is required")
-	}
-	threadRuntime, err := s.rt.NewThreadRuntime(th.ID)
-	if err != nil {
-		return nil, err
-	}
-	if threadRuntime.Toolkit != nil {
-		if _, restoreErr := threadRuntime.Toolkit.RestorePlanFromHistory(history); restoreErr != nil {
-			providers.DebugLogf("restore update_plan for thread %q: %v", th.ID, restoreErr)
-		}
-	}
-	th.mu.Lock()
-	if th.execRuntime == nil {
-		th.execRuntime = threadRuntime
-		th.mu.Unlock()
-		s.subscribeThreadRuntime(th.ID, threadRuntime)
-		return threadRuntime, nil
-	}
-	existing = th.execRuntime
-	th.mu.Unlock()
-	return existing, nil
-}
-
-func (s *Server) subscribeThreadRuntime(threadID string, threadRuntime *runtime.ThreadRuntime) {
-	if threadRuntime == nil || threadRuntime.AgentControl == nil {
-		return
-	}
-	ch := make(chan subagent.Notification, 64)
-	threadRuntime.AgentControl.Subscribe(ch)
-	go s.forwardAgentNotifications(threadID, threadRuntime.AgentControl, ch)
-
-	streamCh := make(chan subagent.StreamNotification, 256)
-	threadRuntime.AgentControl.SubscribeStream(streamCh)
-	go s.forwardAgentStreamNotifications(threadID, threadRuntime.AgentControl, streamCh)
-}
-
-func normalizeTurnStartImages(images []TurnStartImage) ([]providers.InputImage, error) {
-	if len(images) == 0 {
-		return nil, nil
-	}
-	out := make([]providers.InputImage, 0, len(images))
-	for index, image := range images {
-		mediaType := strings.TrimSpace(image.MediaType)
-		data := strings.TrimSpace(image.Data)
-		if data == "" {
-			return nil, fmt.Errorf("image %d data is required", index+1)
-		}
-		var err error
-		mediaType, data, err = normalizeImagePayload(mediaType, data)
-		if err != nil {
-			return nil, fmt.Errorf("image %d: %w", index+1, err)
-		}
-		out = append(out, providers.InputImage{MediaType: mediaType, Data: data})
-	}
-	return out, nil
-}
-
-func normalizeImagePayload(mediaType, data string) (string, string, error) {
-	if strings.HasPrefix(strings.ToLower(data), "data:") {
-		header, payload, ok := strings.Cut(data, ",")
-		if !ok {
-			return "", "", errors.New("invalid data URL")
-		}
-		if !strings.Contains(strings.ToLower(header), ";base64") {
-			return "", "", errors.New("image data URL must be base64")
-		}
-		if mediaType == "" {
-			mediaType = strings.TrimPrefix(strings.Split(header, ";")[0], "data:")
-		}
-		data = strings.TrimSpace(payload)
-	}
-	if mediaType == "" {
-		mediaType = "image/png"
-	}
-	if !strings.HasPrefix(strings.ToLower(mediaType), "image/") {
-		return "", "", fmt.Errorf("unsupported media type %q", mediaType)
-	}
-	if _, err := base64.StdEncoding.DecodeString(data); err != nil {
-		return "", "", fmt.Errorf("invalid base64 data: %w", err)
-	}
-	return mediaType, data, nil
-}
-
-func (s *Server) handleTurnInterrupt(req Request) error {
-	var params TurnInterruptParams
-	if err := decodeParams(req.Params, &params); err != nil {
-		return s.writeResponse(req.ID, nil, err)
-	}
-	th := s.thread(strings.TrimSpace(params.ThreadID))
-	if th == nil {
-		return s.writeResponse(req.ID, nil, fmt.Errorf("thread %q not found", params.ThreadID))
-	}
-	th.mu.Lock()
-	cancel := th.cancel
-	th.mu.Unlock()
-	if cancel == nil {
-		return s.writeResponse(req.ID, nil, errors.New("thread has no running turn"))
-	}
-	cancel()
-	return s.writeResponse(req.ID, OKResult{OK: true}, nil)
-}
-
-func (s *Server) runTurn(ctx context.Context, th *threadState, threadRuntime *runtime.ThreadRuntime, turnID string, history []providers.ChatMessage) {
-	notify := func(method string, params any) {
-		_ = s.writeNotification(method, params)
-	}
-	notifyBatch := func(batch []outboundNotification) {
-		for _, item := range batch {
-			notify(item.method, item.params)
-		}
-	}
-	runner := s.rt.StreamRunner
-	if threadRuntime != nil && threadRuntime.StreamRunner != nil {
-		runner = threadRuntime.StreamRunner
-	}
-	res, err := runner.RunWithCallback(ctx, history, func(ev providers.StreamEvent) {
-		th.mu.Lock()
-		batch := th.applyStreamEventLocked(turnID, ev, time.Now().UTC())
-		th.mu.Unlock()
-		notifyBatch(batch)
-		notify(NotificationTurnEvent, TurnEventNotification{
-			ThreadID: th.ID,
-			TurnID:   turnID,
-			Event:    sanitizeStreamEvent(ev),
-		})
-	})
-
-	now := time.Now().UTC()
-	th.mu.Lock()
-	rewriteHistory := res.HistoryRewritten
-	if res.HistoryRewritten {
-		th.History = append([]providers.ChatMessage(nil), res.NewMessages...)
-	} else {
-		th.History = append(th.History, res.NewMessages...)
-	}
-	var historyErr error
-	if normalized, nerr := providers.NormalizeAndValidateMessages(th.History); nerr != nil {
-		historyErr = nerr
-	} else if !reflect.DeepEqual(normalized, th.History) {
-		th.History = normalized
-		rewriteHistory = true
-	}
-	var persistErr error
-	if historyErr != nil {
-		persistErr = historyErr
-	} else {
-		persistErr = s.persistTurnResultLocked(th, res, rewriteHistory)
-	}
-	status := TurnStatusCompleted
-	if err != nil {
-		status = TurnStatusFailed
-		if errors.Is(err, context.Canceled) {
-			status = TurnStatusInterrupted
-		}
-	}
-	if err == nil && persistErr != nil {
-		err = persistErr
-		status = TurnStatusFailed
-	}
-	var titleHistory []providers.ChatMessage
-	if err == nil {
-		titleHistory = cloneHistory(th.History)
-	}
-	turn := th.completeTurnLocked(turnID, status, err, now)
-	th.mu.Unlock()
-
-	if err != nil {
-		notify(NotificationTurnError, TurnErrorNotification{
-			ThreadID: th.ID,
-			TurnID:   turnID,
-			Error:    err.Error(),
-			Turn:     turn,
-		})
-		s.kickAgentCompletionDrain(th.ID)
-		return
-	}
-	notify(NotificationTurnCompleted, TurnCompletedNotification{
-		ThreadID:     th.ID,
-		Turn:         turn,
-		Content:      res.Content,
-		InputTokens:  res.InputTokens,
-		OutputTokens: res.OutputTokens,
-	})
-	go s.generateThreadTitle(th.ID, titleHistory)
-	s.kickAgentCompletionDrain(th.ID)
-}
-
-func (s *Server) enqueueAgentCompletionTurn(threadID string, msg providers.ChatMessage) {
-	threadID = strings.TrimSpace(threadID)
-	if threadID == "" || (strings.TrimSpace(msg.Content) == "" && len(msg.Images) == 0) {
-		return
-	}
-	th := s.thread(threadID)
-	if th == nil || !canResumeAgentCompletionThread(th) {
-		return
-	}
-	if strings.TrimSpace(msg.Role) == "" {
-		msg.Role = "user"
-	}
-
-	s.agentCompletionMu.Lock()
-	if s.pendingAgentCompletionTurns == nil {
-		s.pendingAgentCompletionTurns = make(map[string][]providers.ChatMessage)
-	}
-	s.pendingAgentCompletionTurns[threadID] = append(s.pendingAgentCompletionTurns[threadID], msg)
-	s.agentCompletionMu.Unlock()
-
-	s.kickAgentCompletionDrain(threadID)
-}
-
-func (s *Server) kickAgentCompletionDrain(threadID string) {
-	threadID = strings.TrimSpace(threadID)
-	if threadID == "" {
-		return
-	}
-
-	s.agentCompletionMu.Lock()
-	if len(s.pendingAgentCompletionTurns[threadID]) == 0 || s.drainingAgentCompletionTurns[threadID] {
-		s.agentCompletionMu.Unlock()
-		return
-	}
-	if s.drainingAgentCompletionTurns == nil {
-		s.drainingAgentCompletionTurns = make(map[string]bool)
-	}
-	s.drainingAgentCompletionTurns[threadID] = true
-	s.agentCompletionMu.Unlock()
-
-	go s.drainAgentCompletionTurns(threadID)
-}
-
-func (s *Server) drainAgentCompletionTurns(threadID string) {
-	th := s.thread(threadID)
-	if th == nil || !canResumeAgentCompletionThread(th) {
-		s.discardPendingAgentCompletionTurns(threadID)
-		s.clearAgentCompletionDrain(threadID)
-		return
-	}
-	if threadIsRunning(th) {
-		s.clearAgentCompletionDrain(threadID)
-		return
-	}
-
-	pending := s.takePendingAgentCompletionTurns(threadID)
-	if len(pending) == 0 {
-		s.clearAgentCompletionDrain(threadID)
-		return
-	}
-
-	started, err := s.startSyntheticTurn(context.Background(), threadID, combineAgentCompletionMessages(pending))
-	if err != nil {
-		providers.DebugLogf("start agent completion turn for thread %q: %v", threadID, err)
-	}
-	requeued := false
-	if !started && err == nil {
-		s.prependPendingAgentCompletionTurns(threadID, pending)
-		requeued = true
-	}
-	s.clearAgentCompletionDrain(threadID)
-	if requeued {
-		s.kickAgentCompletionDrain(threadID)
-	}
-}
-
-func (s *Server) startSyntheticTurn(ctx context.Context, threadID string, userMsg providers.ChatMessage) (bool, error) {
-	threadID = strings.TrimSpace(threadID)
-	if threadID == "" {
-		return false, errors.New("thread_id is required")
-	}
-	if strings.TrimSpace(userMsg.Role) == "" {
-		userMsg.Role = "user"
-	}
-	if strings.TrimSpace(userMsg.Content) == "" && len(userMsg.Images) == 0 {
-		return false, nil
-	}
-
-	th := s.thread(threadID)
-	if th == nil {
-		return false, fmt.Errorf("thread %q not found", threadID)
-	}
-	if !canResumeAgentCompletionThread(th) {
-		return false, nil
-	}
-	threadRuntime, err := s.ensureThreadRuntime(th)
-	if err != nil {
-		return false, err
-	}
-
-	turnID := session.NewID()
-	turnCtx, cancel := context.WithCancel(ctx)
-	now := time.Now().UTC()
-
-	th.mu.Lock()
-	if th.running {
-		th.mu.Unlock()
-		cancel()
-		return false, nil
-	}
-	if th.ReadOnly {
-		th.mu.Unlock()
-		cancel()
-		return false, nil
-	}
-	if err := appendChatMessage(th.MemoryPath, userMsg); err != nil {
-		th.mu.Unlock()
-		cancel()
-		return false, err
-	}
-	history := append([]providers.ChatMessage(nil), th.History...)
-	history = append(history, userMsg)
-	th.History = history
-	th.cancel = cancel
-	turn := th.startTurnLocked(turnID, userMsg, now)
-	th.mu.Unlock()
-
-	_ = s.writeNotification(NotificationTurnStarted, TurnStartedNotification{
-		ThreadID: threadID,
-		Turn:     turn,
-	})
-	go s.runTurn(turnCtx, th, threadRuntime, turnID, history)
-	return true, nil
-}
-
-func canResumeAgentCompletionThread(th *threadState) bool {
-	if th == nil {
-		return false
-	}
-	th.mu.Lock()
-	defer th.mu.Unlock()
-	return !th.ReadOnly
-}
-
-func threadIsRunning(th *threadState) bool {
-	if th == nil {
-		return false
-	}
-	th.mu.Lock()
-	defer th.mu.Unlock()
-	return th.running
-}
-
-func (s *Server) takePendingAgentCompletionTurns(threadID string) []providers.ChatMessage {
-	s.agentCompletionMu.Lock()
-	defer s.agentCompletionMu.Unlock()
-	pending := append([]providers.ChatMessage(nil), s.pendingAgentCompletionTurns[threadID]...)
-	delete(s.pendingAgentCompletionTurns, threadID)
-	return pending
-}
-
-func (s *Server) prependPendingAgentCompletionTurns(threadID string, msgs []providers.ChatMessage) {
-	if len(msgs) == 0 {
-		return
-	}
-	s.agentCompletionMu.Lock()
-	defer s.agentCompletionMu.Unlock()
-	if s.pendingAgentCompletionTurns == nil {
-		s.pendingAgentCompletionTurns = make(map[string][]providers.ChatMessage)
-	}
-	existing := append([]providers.ChatMessage(nil), s.pendingAgentCompletionTurns[threadID]...)
-	s.pendingAgentCompletionTurns[threadID] = append(append([]providers.ChatMessage(nil), msgs...), existing...)
-}
-
-func (s *Server) discardPendingAgentCompletionTurns(threadID string) {
-	s.agentCompletionMu.Lock()
-	defer s.agentCompletionMu.Unlock()
-	delete(s.pendingAgentCompletionTurns, threadID)
-}
-
-func (s *Server) clearAgentCompletionDrain(threadID string) {
-	s.agentCompletionMu.Lock()
-	defer s.agentCompletionMu.Unlock()
-	delete(s.drainingAgentCompletionTurns, threadID)
-}
-
-func combineAgentCompletionMessages(msgs []providers.ChatMessage) providers.ChatMessage {
-	if len(msgs) == 0 {
-		return providers.ChatMessage{Role: "user"}
-	}
-	if len(msgs) == 1 {
-		return msgs[0]
-	}
-	contents := make([]string, 0, len(msgs))
-	for _, msg := range msgs {
-		if content := strings.TrimSpace(msg.Content); content != "" {
-			contents = append(contents, content)
-		}
-	}
-	return providers.ChatMessage{
-		Role:    "user",
-		Content: strings.Join(contents, "\n\n"),
-	}
-}
-
 func (s *Server) thread(id string) *threadState {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 	return s.threads[id]
-}
-
-func (s *Server) persistTurnResultLocked(th *threadState, res agent.LoopResult, rewriteHistory bool) error {
-	if strings.TrimSpace(th.MemoryPath) == "" {
-		return nil
-	}
-	if rewriteHistory {
-		if err := rewriteChatHistory(th.MemoryPath, th.History); err != nil {
-			return err
-		}
-	} else {
-		for _, msg := range res.NewMessages {
-			if err := appendChatMessage(th.MemoryPath, msg); err != nil {
-				return err
-			}
-		}
-	}
-	if err := appendTokenUsage(th.MemoryPath, res.InputTokens, res.OutputTokens); err != nil {
-		return err
-	}
-	return session.UpdateIndex(s.rt.SessionDir, th.ID, persistableMessageCount(th.History), threadPreview(th.History))
 }
 
 type clientResponse struct {
